chore: Update integration tests for Neo 3.9 (#1139) (#1140)
* #1122: first updates (WIP)

* chore: update snapshot repository URLs in DEVELOPMENT.md and add reminders for version updates

* chore: update release workflow to use 'publish' instead of 'publishToSonatype' for snapshot builds

* chore: use minor version bump

* chore: remove neofs module

The neofs module has a dependency to a snapshot version of neofs-api-java which is no longer available. Without it the neofs module nor int-tests which is depending on it can be built.

With int-tests being a crucial part that is required by neow3j we decided to archive all neofs-related code until development of neofs within neow3j is resumed.
Once it is resumed, updating neofs-api-java and publishing a new snapshot will be part of this, but until then, this code will remain in the history (and in another archived branch).

* test: add assert reasons to failing tests in gh action

* test: fix ModuleTest.fastforward with time tolerance

* test: fix expected exception message in neo-express halt test

* test: remove io exception message assertion in neo-express resume test

* test: try neoxp resume test without asserting resume msg

* test: increase tolerance and simplify neoxp resume message

* feat: use neo-node-3.9.1 for privatenet test container

* test: update test for execFeeFactor's updated precision (#1139)

* test: update minor rephrasing of node exception assertion (#1139)

* chore: update read-only integration test for Neo 3.9 (#1139)

* chore (minor): fix misspelling

* test: try to fix neo-express int-test waiting for node resuming

* chore: separate fastForward tests

* chore: use order in fast-forward tests

---------

+++ int-tests/src/test-integration/java/io/neow3j/compiler/PolicyContractIntegrationTest.java
@@ -88,7 +88,11 @@ public class PolicyContractIntegrationTest {
     @Test
     public void setAndGetExecFeeFactor() throws IOException {
         ct.signWithCommitteeAccount();
-        NeoInvokeFunction response = ct.callInvokeFunction(testName, integer(99));
+        // Starting from the Faun hard fork, the setter expects additional precision of 4 decimal places, while
+        // getExecFeeFactor returns the floored value without the additional precision.
+        // 995627 represents 99.5627 with 4 decimal places.
+        // In order to get the full precise value, use getExecPicoFeeFactor().
+        NeoInvokeFunction response = ct.callInvokeFunction(testName, integer(995627));
 
         List<StackItem> res = response.getInvocationResult().getStack().get(0).getList();
         assertThat(res.get(0).getInteger(), is(BigInteger.valueOf(DEFAULT_EXEC_FEE_FACTOR)));
